[x64] d912pxy: Hash entire memory area instead of UINT_MAX bits only

+++ d912pxy/d912pxy_pso_item.cpp
@@ -325,8 +325,8 @@ void d912pxy_pso_item::RCEFixIOBlocksOrdering(char** vsOut, char** psIn, UINT vs
 char* d912pxy_pso_item::RCELinkDerivedCSO(d912pxy_mem_block* src, char* alias)
 {
 	UINT64 derivedUID[2] = {
-		d912pxy::Hash64(d912pxy::MemoryArea(src[0].ptr(), (UINT)src[0].size())).value,
-		d912pxy::Hash64(d912pxy::MemoryArea(src[1].ptr(), (UINT)src[1].size())).value
+		d912pxy::Hash64(d912pxy::MemoryArea(src[0].ptr(), src[0].size())).value,
+		d912pxy::Hash64(d912pxy::MemoryArea(src[1].ptr(), src[1].size())).value
 	};
 
 	char derivedName[255];
